[REF] extension: new architecture
The 3 main components (page bridge, content and service worker)
each have their own encapsulated modules instead of being
big monolythic files
