+++ extension/manifest.json
@@ -2,7 +2,7 @@
     "manifest_version": 3,
     "name": "Discuss Companion Bridge",
     "description": "Bridge extension to connect to the Discuss Companion app",
-    "version": "0.10.0",
+    "version": "0.10.1",
     "homepage_url": "https://github.com/ThanhDodeurOdoo/discuss-companion",
     "permissions": [
         "storage",
